cmd/{containerboot,k8s-operator},kube/kubetypes: unadvertise ingress services on shutdown
Add a PreStop lifecycle hook to the ProxyGroup ingress spec so that we
can explicitly notify control that the service is being unadvertised and
prompt it to update the netmap for clients faster. Without this, control
just treats the device as temporarily offline and allows it a grace period
before no longer considering it part of the service.

Updates tailscale/corp#24795

Change-Id: I0a9a4fe7a5395ca76135ceead05cbc3ee32b3d3c
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>