cmd/{containerboot,k8s-operator},kube/kubetypes: unadvertise ingress services on shutdown
Ensure no services are advertised as part of shutting down tailscaled.
Prefs are only edited if services are currently advertised, and updating
prefs ensures that control is proactively notified to prompt it to update
the netmap for clients faster. Without this, control just treats the device
as temporarily offline and allows it a grace period before no longer
considering it part of the service.

Note that editing prefs will trigger a synchronous write to the state
Secret, so it may fail to persist state if the ProxyGroup is getting
scaled down and therefore has its RBAC deleted at the same time, but that
failure occurs after control has been notified, and the state Secret is
about to get deleted anyway, so the only negative side effect is a harmless
error log during shutdown.

Note that the first version of this used a PreStop lifecycle hook, but
that only supports GET methods and we need the shutdown to trigger side
effects (updating prefs) so it didn't seem appropriate to expose that
functionality on a GET endpoint that's accessible on the k8s network.

Updates tailscale/corp#24795

Change-Id: I0a9a4fe7a5395ca76135ceead05cbc3ee32b3d3c
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>