drive: fix index out of bounds when parsing request local paths
Fix the index out of bound panic when a request is made to the local
fileserver mux with a valid secret-token, but missing share name.

Example error:
http: panic serving 127.0.0.1:40974: runtime error: slice bounds out of range [2:1]

Additionally, we documentation the edge case behavior of utilities that
this fileserver mux depends on.

Signed-off-by: Craig Hesling <craig@hesling.com>