break up health watchers
Some watchers need to know about every warnable change (E.g. to report to
control), some send the UI the full list of warnable changes on any change.
Let's make a separate type for the latter so we have a handy way to send all
UnhealthyStates to the UI when ControlHealth messages change.