{control/controlhttp,derp/derphttp}: restore websocket dependency tests
This is a partial revert of commit 6d6f69e7358f.

Instead of ruling out the websocket dependency for all build targets, do so
specifically for iOS, where the concern is to avoid unduly increasing the size
of the target binary.  Since there is no way on iOS to get access to the debug
services that use websockets, we already disable the dependendency on iOS
anyway.

Updates #15160

Change-Id: I2fb3d9aafaaca8d1037fe9dea3462e85b8bc1e27
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>