ipn/ipnlocal: move resetDialPlan to after cc.Logout() returns
We shouldn't be resetting the current dial plan until cc.Logout() has returned,
so we move the resetDialPlan() call accordingly.

We also update a comment that become inaccurate sometime since 2020.

Finally, we add a resetDialPlan() call when the current profile is deleted.
It results in a switch to a new, empty profile and resets the state machine,
so the dial plan should be reset as well.

Updates #cleanup
Updates #11938

Signed-off-by: Nick Khyl <nickk@tailscale.com>