net/dns/resolvd: prevent overwriting our backup resolv.conf (#12360)
When SetDNS is called multiple times, such as on start up, the current
implementation will create backupConf, and then subsequent calls will
overwrite our backup with the last generated resolv.conf file. Guard
against this by only copying the backupConf the first time.

Write the sentinel comments that allow us to check whether we "own" the
resolv.conf as resolvconffile does. Also do some clean up of the resolvd
nameserver entries and whitespace, to enable a follow up commit to
transition to writing the config with resolvconffile instead of open
coding it.

Fixes #12360

Signed-off-by: Kyle McMartin <jkkm@jkkm.org>