cmd/containerboot: load auth key from file if TS_AUTHKEY_FILE is set
This changes adds a new environment variable `TS_AUTHKEY_FILE`, that
can be used instead of `TS_AUTHKEY` to specify the auth key to use for
login. TS_AUTHKEY takes precedence over TS_AUTHKEY_FILE, if both are
set.

This is useful for cases where the auth key is stored in a docker secret
or similar, and makes it possible to avoid storing the auth key in a
docker compose file.

Fixes <https://github.com/tailscale/tailscale/issues/14210>

Signed-off-by: Josh McKinney <joshka@users.noreply.github.com>