feature/taildrop: do not use m.opts.Dir for Android
In Android, we are prompting the user to select a Taildrop directory when they first receive a Taildrop: we block writes on Taildrop dir selection. This means that we cannot use Dir inside managerOptions, since the http request would not get the new Taildrop extension. This PR removes, in the Android case, the reliance on m.opts.Dir, and instead has FileOps hold the correct directory.
This also cleans up some of the redundant logic around checking if FileOps should be used: we remove PutMode and instead rely on DirectFileMode and FileOps to determine whether this is a direct transfer, direct transfer using Android, or indirect (stages first).
Follow-up PR will use SAF for resuming partial Taildrops.

Updates tailscale/corp#29211

Signed-off-by: kari-ts <kari@tailscale.com>